% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affTr.R
\name{affTr}
\alias{affTr}
\title{This function affine transforms images to create template for each group or species.}
\usage{
affTr(img_path,img_ID, target, out_path)
}
\arguments{
\item{img_path}{Folder path for raw images.}

\item{img_ID}{A matrix containing the names of group identifiers in the first column and
the sequence number of the image in that group to be affine transformed.}

\item{target}{Name of the target group or species name (should be one of the names in the
first column of img_ID)}

\item{out_path}{Folder path for template images.}
}
\value{
None. Affine-transformed images are saved in the user-defined folder.
}
\description{
Affine transforms images to a template for each group/species.
}
\details{
This function affine-transforms raw .jpg images to a template for each
group/species using utilities of R package "RNiftyReg". These transformed images
will be used further in the function colData for colour pattern extraction.
}
\examples{
\dontrun{
affTr(img_path = "img_folder/", img_ID=colour_sampling,target= "Tiru_sept",
out_path = "transformed_folder/")
}

}
