% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binImg.R
\name{binImg}
\alias{binImg}
\title{Creating square binary images for outline analysis.}
\usage{
binImg(img_path, thresh, dimn, out_path)
}
\arguments{
\item{img_path}{The path of the raw image files.}

\item{thresh}{The thresholding cutoff (ranges from 0 to 1). defaults to 0.9. Images with a darker foreground require
lower threshold values to efficiently segregate the background and foreground.}

\item{dimn}{Number of pixels (length and width) for image resizing. Defaults to 100.}

\item{out_path}{Folder path for saving the binarized images.}
}
\value{
None. Transformed and resized binary images are directly saved in the user-defined folder.
}
\description{
Transforming raw images into binary images (black foreground, white background) for the analysis of shape
outlines.
}
\details{
A robust function that thresholds .jpg images of specimens photographed against a lighter background (close to
 white) and transforms them into a square blank-and-white image that is resized to user-defined dimensions. This
 transformed images are directly saved to a user-defined folder.
}
\examples{
\dontrun{
binImg(img_path = "img_folder/", thresh = 0.97, dimn = 100, out_path = "test/")
}

}
