% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BLU_IDList}
\alias{BLU_IDList}
\title{BLU_IDList}
\format{
A list of group names including the blue colour pattern extracted from
    dorsal forewings of six butterfly species using \code{colData}.
}
\usage{
BLU_IDList
}
\description{
List of group names as returned by \code{colData}.
}
\examples{
\dontrun{
data(BLU_IDList)
summary(BLU_IDList)
}
}
\keyword{datasets}
