% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BLU_mean}
\alias{BLU_mean}
\title{BLU_mean}
\format{
A dataframe containing  linearized proportional data of the blue colour pattern extracted from
   dorsal forewings of six butterfly species using \code{colData}.
}
\usage{
BLU_mean
}
\description{
A linearized dataframe of colour pattern data returned by \code{colData}.
}
\examples{
\dontrun{
data(BLU_mean)
summary(BLU_mean)
}
}
\keyword{datasets}
