% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BLU_RasterStacks}
\alias{BLU_RasterStacks}
\title{BLU_RasterStacks}
\format{
A list of RasterLayers including the blue color pattern extracted from
   dorsal forewings of  six butterfly species using \code{colData}.
}
\usage{
BLU_RasterStacks
}
\description{
List of RasterLayers as returned by \code{colData}.
}
\examples{
\dontrun{
data(BLU_RasterStacks)
summary(BLU_RasterStacks)
}
}
\keyword{datasets}
