% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cIndex.R
\name{cIndex}
\alias{cIndex}
\title{Estimating convergence among multiple species/groups.}
\usage{
cIndex(x , y , z , dimn , nsim)
}
\arguments{
\item{x}{A dataframe containing PCA results.}

\item{y}{A phylogenetic tree of class phylo.}

\item{z}{A list of groups containing species names among which convergence indices are to be estimated.}

\item{dimn}{Number of PC axes to be used when calculating C indices.}

\item{nsim}{Number of rounds of simulation for estimating the null distribution of C indices. Defaults to 100.}
}
\value{
A list object containing estimated C indices, their p-values, the simulated null distributions, and
respective cutoffs.
}
\description{
Estimating C indices (Stayton, 2015) to examine the extent of convergence in the morphospace.
}
\details{
This function is based on the 'convevol' package (Stayton, 2015) and estimates convergence indices C1–C4
(all of these range from 0 to 1). Values closer to 0 indicate low convergence, and values closer to 1 indicate higher
convergence. This function also simulates the distributions of C indices using the Brownian motion model of evolution
and estimates p-values for each estimated C-index.
}
\examples{
\dontrun{
regimes <- list(rgm1 = c("sp1", "sp2"), rgm2 = c("sp3", "sp1"))
cIndex(x = pc.shape, y = tree, z = regimes, dimn = 4, nsim = 100)
}

}
