% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleqEFA.R
\name{cleqEFA}
\alias{cleqEFA}
\title{Elliptical Fourier analysis on closed contours.}
\usage{
cleqEFA(img_path, out_path)
}
\arguments{
\item{img_path}{Folder path for binary image files (same as the output folder for the binImg function).}

\item{out_path}{Folder path for output images and final EFA descriptors.}
}
\value{
None. Images showing the contour stacks, harmonic power, results of elliptical analysis, and fitted ellipses
(intermediate results). The EFA descriptors and meanshapes for each species/group are saved in the output folder.
}
\description{
This function applies Fourier transform on outlines.
}
\details{
This function uses utilities provided in the package "Momocs" and performs
Elliptical Fourier analysis on two-dimensional closed contours. This is an interactive
function that scales, centres, and aligns closed outlines from a list of binary images.
}
\examples{
\dontrun{
cleqEFA(img_path = "img_folder/", out_path = "output_folder/")
}

}
