% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{col_concat}
\alias{col_concat}
\title{col_concat}
\format{
A concatenated linearized dataframe of colour pattern data of multiple colours extracted from dorsal
forewings of six butterfly species using \code{concat}.
}
\usage{
col_concat
}
\description{
A linearized dataframe of colour pattern data returned by \code{concat}.
}
\examples{
\dontrun{
data(col_concat)
summary(col_concat)
}
}
\keyword{datasets}
