% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{col_index}
\alias{col_index}
\title{col_index}
\format{
A matrix containing the names of group identifiers in the first column and
        the sequence number of the image in that group to be used as the template for alignment. It is used in \code{affTr},
        \code{colData}.
}
\usage{
col_index
}
\description{
Group identifiers and the sequence number of a template image for each group.
}
\examples{
\dontrun{
data(col_index)
summary(col_index)
}
}
\keyword{datasets}
