% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{col_pcscores}
\alias{col_pcscores}
\title{col_pcscores}
\format{
A dataframe containing PC scores obtained from \code{\link[phytools]{phyl.pca}} on concatenated colour
data returned by \code{concat}.
}
\usage{
col_pcscores
}
\description{
A dataframe containing PC scores.
}
\examples{
\dontrun{
data(col_pcscores)
summary(col_pcscores)
}
}
\keyword{datasets}
