% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colComp.R
\name{colComp}
\alias{colComp}
\title{Comparative visualization of colour pattern differences.}
\usage{
colComp(gr1, gr2, rstrstcks, ID, outline_path, bcg_col, font_col)
}
\arguments{
\item{gr1}{A vector of species names.}

\item{gr2}{Another vector of species names (gr2 could be similar to gr 1 or different).}

\item{rstrstcks}{A list object containing rasterstacks for each species as obtained from colData.}

\item{ID}{A raster ID containing specimen names as saved from colData.}

\item{outline_path}{A path containing the text file (.txt) containing outline coordinates and a binary
image of a cartton template on which heatmaps will be plotted.}

\item{bcg_col}{Background colour to be displayed in the graphics (default="black").}

\item{font_col}{Text colour to be displayed on the graphics (default="white").}
}
\value{
None. a multipanel graphics showing heatmaps as extent of colour pattern variations within and between species
groups.
}
\description{
Comparative visualization of the extent of differences between homologous colour pattern
elements among various groups.
}
\details{
The function  estimates differences between mean colour patterns of various species and displays
a heatmap showing the colour pattern variations within and between species groups. This function is based
on the  utilities of R package 'patternize'.
}
\examples{
\dontrun{
gr1<-c("Idea_mala", "Para_agle", "Para_nilg", "Tiru_limn", "Tiru_sept")
gr2<-c("Pare_ceym", "Pare_ceyf", "Pare_hipm", "Pare_hipf", "Papl_cldm", "Papl_cldf")
colComp(gr1, gr2, rstrstcks=Mel_rasterstacks, ID=Mel_IDlist, outline_path="/outline/",
bcg_col="black", font_col="white")
}

}
