% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colData.R
\name{colData}
\alias{colData}
\title{Data extraction for colour analysis.}
\usage{
colData(sp_index, img_folder, ext, target, sampletype, resampleFactor, out_folder)
}
\arguments{
\item{sp_index}{A list indicating the species to be sampled for each colour.}

\item{img_folder}{A folder path for cleaned original and affine-transformed images.}

\item{ext}{File extension for all image files from which data need to be extracted.}

\item{target}{A matrix containing the names of group identifiers in the first column and
the sequence number of the image in that group to be used as the template for alignment.}

\item{sampletype}{Same as in \code{\link[patternize]{sampleRGB}}.}

\item{resampleFactor}{Same as in \code{\link[patternize]{sampleRGB}}.}

\item{out_folder}{A folder path for all outputs.}
}
\value{
None. The function directly saves the proportional data matrix, the sampled RGB values and offsets,
stacked images, and the species-level rasterStacks to a user-defined folder.
}
\description{
This function extracts a proportional data matrix representing the species-level average presence/absence
of a colour in each pixel of a square image.
}
\details{
Colours (RGB values) are sampled from a target image, and an offset is defined to calculate the range of RGB
values to be sampled from all images of a species. All images of a species are aligned (stacked) against the respective
target image for that species. All pixels containing RGB values within the user-defined range are highlighted.
This is converted into individual-level binary matrices, which are then averaged into a species-level proportional matrix.
}
\examples{
\dontrun{
colData(sp_index=sp_index, img_folder = "img_folder", ext = ".jpg",
target = colour_sampling, sampletype = "point", resampleFactor = 1, out_folder= "output_folder/")
}

}
