% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat.R
\name{concat}
\alias{concat}
\title{Concatenation of large colour data files}
\usage{
concat(file_names)
}
\arguments{
\item{file_names}{A list of file names containing proportional matrices for each colour as obtained
from function colData.}
}
\value{
a large dataframe
}
\description{
Concatenated large data frames into a wide conjugate dataframe.
}
\details{
This function concatenates the wide dataframes of proportional matrices of each colour range into a large
dataframe.
}
\examples{
\dontrun{
file_names<-list(MEL_mean, BLU_mean, RED_mean)
concat(file_names)
}

}
