% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MEL_IDList}
\alias{MEL_IDList}
\title{MEL_IDList}
\format{
A list of group names including the melanistic colour pattern extracted from
    dorsal forewings of six butterfly species using \code{colData}.
}
\usage{
MEL_IDList
}
\description{
List of group names as returned by \code{colData}.
}
\examples{
\dontrun{
data(MEL_IDList)
summary(MEL_IDList)
}
}
\keyword{datasets}
