% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MEL_mean}
\alias{MEL_mean}
\title{MEL_mean}
\format{
A dataframe containing  linearized proportional data of the melanistic colour pattern extracted from
   dorsal forewings of six butterfly species using \code{colData}.
}
\usage{
MEL_mean
}
\description{
A linearized dataframe of colour pattern data returned by \code{colData}.
}
\examples{
\dontrun{
data(MEL_mean)
summary(MEL_mean)
}
}
\keyword{datasets}
