% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MEL_RasterStacks}
\alias{MEL_RasterStacks}
\title{MEL_RasterStacks}
\format{
A list of RasterLayers including the melanistic color pattern extracted from
   dorsal forewings of  six butterfly species using \code{colData}.
}
\usage{
MEL_RasterStacks
}
\description{
List of RasterLayers as returned by \code{colData}.
}
\examples{
\dontrun{
data(MEL_RasterStacks)
summary(MEL_RasterStacks)
}
}
\keyword{datasets}
