% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mnshp_efa}
\alias{mnshp_efa}
\title{mnshp_efa}
\format{
A dataframe containing elliptical Fourier descriptors for each harmonics explaining the mean shape of each group.
}
\usage{
mnshp_efa
}
\description{
Fourier descriptors obtained from \code{cleqEFA} and used by
   \code{dataPrep}.
}
\examples{
\dontrun{
data(mnshp_efa)
summary(mnshp_efa)
}
}
\keyword{datasets}
