% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mnshps}
\alias{mnshps}
\title{mnshps}
\format{
A list of x, y coordinates of forewing mean shape of six butterfly species.
}
\usage{
mnshps
}
\description{
Coordinates of mean shapes returned by \code{cleqEFA} and used by
   \code{shpComp}.
}
\examples{
\dontrun{
data(mnshps)
summary(mnshps)
}
}
\keyword{datasets}
