% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcModels.R
\name{pcModels}
\alias{pcModels}
\title{Evolutionary model fitting}
\usage{
pcModels(phy, dat)
}
\arguments{
\item{phy}{A phylogenetic tree of class phylo.}

\item{dat}{A vector or matrix containing dependent variable/s.}
}
\value{
A list of the estimated statistical parameters for each models fitted to each PC.
}
\description{
Evolutionary model fitting
}
\details{
This function fits six macroevolutionary models to a phylogenetic tree: Brownian motion (BM), early burst (EB),
Ornstein–Uhlenbeck (OU), lambda, kappa, and delta.  The function returns the estimated statistical parameters for each
models fitted to each PC.
}
\examples{
\dontrun{
evo_model<-pcModels(sample.tree.Tiru, PCA_col)
}

}
