% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rateEvol.R
\name{rateEvol}
\alias{rateEvol}
\title{Estimate rate of trait evolution}
\usage{
rateEvol(traitPCA, dimn, tree, ntip, intr, tiplist)
}
\arguments{
\item{traitPCA}{A list containing PCA elements obtained from dataPrep implemented on different traits.}

\item{dimn}{Number of PC dimension to be used to estimate evolutionary rate.}

\item{tree}{A phylogenetic tree of class phylo.}

\item{ntip}{Number of tips for which tree and PCA data has been provided.}

\item{intr}{A logical argument, whether tips were introduced manually and rates need to be calculated for those
introduced tips separately (defaults to FALSE).}

\item{tiplist}{A list file containing two sets of tip names that will be used to separately calculate rates for
the introduced branches (defaults to NULL).}
}
\value{
A dataframe containing rates as coefficients of ridge regression for all the branches of the tree.
}
\description{
Estimate rate of trait evolution using a ridge regression method.
}
\details{
This function based on the utilities of R package 'RRphylo' performs ridge regression and simultaneously
estimates rate of  multiple trait evolution along all the branches of a phylogenetic tree.Rates of evolution
can also be estimated for incompletely resolved (polytomic) phylogenies or phylogenies with artificially
introduced branches (e.g., individual branches for conspecific forms or sexes). In such cases, the
'rateEvol()' function considers the introduced branches one at a time (while omitting its sister branches)
and calculates the rate of evolution from the node that represents species divergence. This process is repeated
as many times as the number of introduced branches. For rate estimation with introduced branches, the function
requires a user input in the form of a list  with two nested list objects. The first list object contains the tip
labels of introduced branches (for which the rate of evolution will be estimated in each iteration), and the second
contains the tip labels of introduced sister branches (that need to be omitted in each iteration).
}
\examples{
\dontrun{
traitPCA<-list(PCA_shp, PCA_col)
tree<-sample.tree.Tiru
tiplist<-list(c("Papl_cldm","Papl_cldf", "Pare_ceym","Pare_ceyf","Pare_hipm","Pare_hipf"),
c("Papl_cldf","Papl_cldm", "Pare_ceyf","Pare_ceym","Pare_hipf","Pare_hipm"))
rateEvol(traitPCA, dimn=4, tree, ntip=29, intr=T, tiplist)
}

}
