% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RED_IDList}
\alias{RED_IDList}
\title{RED_IDList}
\format{
A list of group names including the red colour pattern extracted from
    dorsal forewings of six butterfly species using \code{colData}.
}
\usage{
RED_IDList
}
\description{
List of group names as returned by \code{colData}.
}
\examples{
\dontrun{
data(RED_IDList)
summary(RED_IDList)
}
}
\keyword{datasets}
