% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RED_RasterStacks}
\alias{RED_RasterStacks}
\title{RED_RasterStacks}
\format{
A list of RasterLayers including the red color pattern extracted from
   dorsal forewings of  six butterfly species using \code{colData}.
}
\usage{
RED_RasterStacks
}
\description{
List of RasterLayers as returned by \code{colData}.
}
\examples{
\dontrun{
data(RED_RasterStacks)
summary(RED_RasterStacks)
}
}
\keyword{datasets}
