% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shp_pcscores}
\alias{shp_pcscores}
\title{shp_pcscores}
\format{
A dataframe containing PC scores obtained from \code{\link[phytools]{phyl.pca}} on Fourier descriptors of
shape data returned by \code{cleqEFA}
}
\usage{
shp_pcscores
}
\description{
A dataframe containing PC scores.
}
\examples{
\dontrun{
data(shp_pcscores)
summary(shp_pcscores)
}
}
\keyword{datasets}
