% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shpComp.R
\name{shpComp}
\alias{shpComp}
\title{Comparative visualization of shape differences}
\usage{
shpComp(gr1, gr2, mnshp)
}
\arguments{
\item{gr1}{A vector of species names .}

\item{gr2}{Another vector of species names (Can be the same as gr1).}

\item{mnshp}{A meanshape file obtained from \code{\link{cleqEFA}}.}
}
\value{
A multi-panel graphics showing the heatmaps of shape deformation between species/groups.
}
\description{
Comparative visualization of the extent of deformation between the two-dimensional contours of various
species/groups.
}
\details{
This function is based on the  'Momocs' package. It estimates the differences in mean shapes of various
species and displays a heatmap showing the extent of deformation.
}
\examples{
\dontrun{
gr1<-c("Idea_mala", "Para_agle", "Para_nilg", "Tiru_limn", "Tiru_sept")
gr2<-c("Pare_ceym", "Pare_ceyf", "Pare_hipm", "Pare_hipf", "Papl_cldm", "Papl_cldf")
data("meanshape")
shpComp(gr1, gr2, mnshp=meanshape)
}

}
