% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wg_tree}
\alias{wg_tree}
\title{wg_tree}
\format{
A phylogenetic tree of class "phylo" of 29 species of butterflies from three families (Nymphalidae,
Pieridae and Papilionidae).
}
\usage{
wg_tree
}
\description{
A "phylo" object.
}
\examples{
\dontrun{
data(wg_tree)
summary(wg_tree)
}
}
\keyword{datasets}
